+++ packages/playground/src/index.tsx
@@ -30,15 +30,35 @@ import { ComponentsProvider, Flex, Icon } from '@looker/components'
 const App = () => {
   return (
     <ComponentsProvider>
-      <Flex alignItems="center" justifyContent="center">
+      <Flex alignItems="center" justifyContent="space-evenly">
+        <Icon
+          artwork={
+            <svg height="24" width="24">
+              <circle cx="12" cy="12" r="12" stroke="black" fill="red" />
+            </svg>
+          }
+          size="xsmall"
+        />
+        <Icon
+          artwork={
+            <svg
+              width="24"
+              height="24"
+              viewBox="0 0 24 24"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M12 2C6.48 2 2 6.48 2 12C2 17.52 6.48 22 12 22C17.52 22 22 17.52 22 12C22 6.48 17.52 2 12 2ZM13 17H11V15H13V17ZM13 13H11V7H13V13Z"
+                fill="#1C2125"
+              />
+            </svg>
+          }
+          name="GearOutline"
+          size="large"
+        />
         <Icon name="GearOutline" size="xxsmall" />
-        <Icon name="GearOutline" size="xsmall" />
-        <Icon name="GearOutline" size="small" />
-        <Icon name="GearOutline" size="medium" />
-        <Icon name="GearOutline" />
-        <Icon name="GearOutline" size="large" />
-        <Icon name="GearOutline" size="78px" />
-        <Icon name="GearOutline" size="90px" />
+        <Icon size="xxsmall" />
       </Flex>
     </ComponentsProvider>
   )